Split out bot tests
